refactor: move each eval type into separate modules
Instead of one big eval.rs, there is now an `eval/` directory with
individual modules for number operations, relational operations,
logical operations, conditionals and functions (lambdas).

Everything in the internal modules are re-exported by eval, so the API
does not change.

Also ran cargo fmt on them.
